Two significant changes.  First, CyTableEntry SUID is now a Long instead of a long.  This avoids a bunch of unnecessary Long.valueOf(long), which takes time and memory.  Second, this change replaces ArrayGraph and ArraySubGraph with SimpleNetwork, DefaultTableNetwork, CyRootNetworkImpl, and CySubNetworkImpl.  This small hierarchy greatly simplifies the implementations of CyNetworks and organizes the code in a way that is easier to understand.  Oh, and the implementations are faster and more time efficient.  Hooray.

+++ model-api/src/test/java/org/cytoscape/model/DummyGraphObject.java
@@ -38,15 +38,15 @@ package org.cytoscape.model;
 
 public class DummyGraphObject implements CyTableEntry {
 	
-	private static int suidBase = 0;
+	private static long suidBase = 0;
 
-	private final long suid;
+	private final Long suid;
 
 	public DummyGraphObject() {
 		suid = suidBase++;
 	}
 
-	public long getSUID() {
+	public Long getSUID() {
 		return suid;
 	}
